+++ src/audio_metadata/structures.py
@@ -62,3 +62,8 @@ class ListMixin(UserList):
 
 	def __repr__(self):
 		return f"<{self.__class__.__name__} ({len(self)} {self.item_label})>"
+
+	@property
+	def items(self):
+		return self.data
+